Migrate deviceprofile to typescript
